tickets/src/models/__test__/ticket.test.ts
import { Ticket } from '../ticket';

it('implement optimistic concurrency control', async () => {
  // create an instance of an ticket
  const ticket = Ticket.build({
    title: 'concert',
    price: 5,
    userId: '1234',
  });

  // save the ticket to the database
  await ticket.save(); // when we save it the mongoose-update-if-current plugin assign a version to it

  // fetch the ticket twice
  const firstInstance = await Ticket.findById(ticket.id);
  const secondInstance = await Ticket.findById(ticket.id);

  // make two seperate changes to the ticket we fetched
  firstInstance?.set({ price: 10 });
  secondInstance?.set({ price: 15 });

  // save the first fetched ticket
  await firstInstance?.save();

  // save the second fetched ticket and expect an error
  try {
    await secondInstance!.save();
  } catch (err) {
    return;
  }

  throw new Error('Should not reach this point');
});

it('increments the version number on multiple saves', async () => {
  const ticket = Ticket.build({
    title: 'concert',
    price: 20,
    userId: '1234',
  });
  await ticket.save();
  expect(ticket.version).toEqual(0);

  await ticket.save();
  expect(ticket.version).toEqual(1);

  await ticket.save();
  expect(ticket.version).toEqual(2);
});
